Final fixes

+++ src/items/items_definitions.hpp
@@ -491,7 +491,7 @@ enum ItemParseAttributes_t {
 
 struct ImbuementInfo {
 	Imbuement *imbuement;
-	uint64_t duration = 0;
+	uint32_t duration = 0;
 };
 
 #endif  // SRC_ITEMS_ITEMS_DEFINITIONS_HPP_
